channel samples() handles time vs distance and relative/absolute
I had been filtering these externally, but decided to just jam all the
complexity into the function itself, so the caller can ask for the data
in readily consumeable way.

This is more delicate than I like, and the defaults are more for
interactive use (distance and relative) than not, so the unittests have
to pass flags to get old style data.